get documentation generator to work
